Rename maven.py into maven_utils.py
